LayoutTests/media/modern-media-controls/fullscreen-button/fullscreen-button.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>FullscreenButton</code> class.");

const enterFullscreenButton = new FullscreenButton({ layoutTraits: LayoutTraits.macOS });

shouldBeEqualToString("enterFullscreenButton.element.localName", "button");
shouldBeTrue('enterFullscreenButton.element.classList.contains("icon")');
shouldBeTrue('enterFullscreenButton.element.classList.contains("fullscreen")');
shouldBeEqualToString("enterFullscreenButton.iconName", `${Icons.EnterFullscreen}`);

const iOSEnterFullscreenButton = new FullscreenButton({ layoutTraits: LayoutTraits.iOS });

const exitFullscreenButton = new FullscreenButton({ layoutTraits: LayoutTraits.macOS | LayoutTraits.Fullscreen });
shouldBeEqualToString("exitFullscreenButton.iconName", `${Icons.ExitFullscreen}`);

scheduler.frameDidFire = function()
{
    if ([enterFullscreenButton, iOSEnterFullscreenButton, exitFullscreenButton].some(button => button.width == 0))
        return;

    shouldBeTrue('enterFullscreenButton.element.style.webkitMaskImage.includes("macOS/enter-fullscreen@")');
    shouldBeTrue('iOSEnterFullscreenButton.element.style.webkitMaskImage.includes("iOS/enter-fullscreen@")');
    shouldBeTrue('exitFullscreenButton.element.style.webkitMaskImage.includes("macOS/exit-fullscreen@")');
    finishJSTest();
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
